fix avatar functionality in the user update method

+++ app/Http/Controllers/UserController.php
@@ -52,14 +52,18 @@ class UserController extends Controller
         $user = User::findOrFail($id);
 
         Gate::authorize('manage-profile', $user);
-
-        if(empty($request->validated()['avatar_link'])) {
+        
+        if(empty($request->validated()['avatar_link']) && isset($user->avatar_path)) {
             $this->fileService->destroy($user->avatar_path);
+            $user->avatar_path = null;
+            $user->save();
         }
 
-        $avatarPath = $this->fileService->getFilePath($request->validated()['avatar_link']);
-        $userData = array_merge($request->validated(), ['avatar_path' => $avatarPath]);
-        $user->update($userData);
+        if($request->validated()['avatar_link']) {
+            $avatarPath = $this->fileService->getFilePath($request->validated()['avatar_link']);
+            $userData = array_merge($request->validated(), ['avatar_path' => $avatarPath]);
+            $user->update($userData);
+        }
 
         UserUpdated::dispatch($user);
